// lib/api/registry.ts
export const API_DOMAINS = [
  'system',
  'user',
  'admin',
  'dashboard',
  'zones',
  'records',
  'dnssec',
  'cache',
  'allowed',
  'blocked',
  'apps',
  'dhcp',
  'logs',
  'settings',
  'dnsClient',
] as const

export type ApiDomain = (typeof API_DOMAINS)[number]